refactor: rename genui coercion helpers to reflect broader usage
