+++ src/components/chat/genui/components/ComparisonTable.tsx
@@ -1,5 +1,5 @@
 import { Check, Minus, X } from 'lucide-react'
-import { coerceArray } from './chart-utils'
+import { coerceArray } from './input-coercion'
 
 interface Feature {
   label: string
